move JSON runtime to src
